code refactor

+++ database/context.go
@@ -44,17 +44,17 @@ func CountDocuments(filter bson.M) int64 {
 	return cantMutants
 }
 
-func calculateRatio(num1 int, num2 int) float32 {
+func calculateRatio(num1 float32, num2 float32) float32 {
 	if num1 > 0 && num2 > 0 {
-		return float32(num1 / num2)
+		return num1 / num2
 	} else {
 		return 0
 	}
 }
 
 func StatisticsBD() status {
-	cantMutants := int(CountDocuments(bson.M{"ismutant": true}))
-	cantNoMutants := int(CountDocuments(bson.M{"ismutant": false}))
+	cantMutants := float32(CountDocuments(bson.M{"ismutant": true}))
+	cantNoMutants := float32(CountDocuments(bson.M{"ismutant": false}))
 
 	return status{
 		CountMutant: cantMutants,
