+++ database/models.go
@@ -1,7 +1,7 @@
 package database
 
 type status struct {
-	CountMutant int     `json:"count_mutant_dna"`
-	CountHuman  int     `json:"count_human_dna"`
+	CountMutant float32 `json:"count_mutant_dna"`
+	CountHuman  float32 `json:"count_human_dna"`
 	Ratio       float32 `json:"ratio"`
 }
